feat: architectural review 2026 and initial refactoring
- Added detailed architectural review findings in `docs/ARCHITECTURAL_REVIEW_2026_RESULTS.md`.
- Refactored `content/main.js` to decouple console filtering and global event delegation.
- Extracted logic into `content/core/console-filter.js` and `content/core/global-events.js`.
- Modernized `ThreeEarthSystem` with `ResizeObserver` for better performance and responsiveness.
- Improved import path consistency using relative paths in components.
- Identified discrepancies between documentation and current repository state (missing Vite/Vitest).
