+++ docs/ARCHITECTURAL_REVIEW_2026_RESULTS.md
@@ -1,0 +1,49 @@
+# Architectural Review Results 2026
+
+**Date:** February 6, 2026
+**Reviewer:** Jules (AI Assistant)
+**Scope:** Repository `iweb`
+
+## 1. Overview
+
+The repository demonstrates a high level of maturity and technical excellence. The codebase is modular, type-safe (via JSDoc), and optimized for performance.
+
+## 2. Key Findings
+
+### 2.1 Technical Strengths
+- **Type Safety:** 100% type safety achieved using JSDoc and `@ts-check`. The `tsc --noEmit` check passes with zero errors.
+- **Modularity:** Components like `RobotCompanion` and `ThreeEarthSystem` are well-isolated and follow clear architectural patterns.
+- **Performance:** Advanced techniques such as viewport-based pausing for 3D rendering, adaptive quality levels, and modern image formats (AVIF/WebP) are implemented.
+- **Separation of Concerns:** Clear distinction between core services, UI components, and page content.
+- **Cloudflare Integration:** Efficient use of Cloudflare Workers and middleware for dynamic content injection.
+
+### 2.2 Discrepancies & Issues
+- **Missing Build System:** Documentation (`PROJECT_STATUS.md`, `REVIEW_REPORT_2026.md`) references a Vite-based build system that is currently not present in the root directory. There is no `vite.config.js` and no `build` script in `package.json`.
+- **Missing Tests:** Documentation mentions 38 passing unit tests, but no test files (`*.test.js`) were found in the repository.
+- **CSS Waterfalls:** The current modular CSS loading strategy (using `@import` and dynamic `link` injection) leads to multiple waterfall requests, which could be optimized via bundling.
+- **Bloated `main.js`:** The application entry point handles multiple responsibilities including bootstrap, event delegation, service worker registration, and loading logic.
+
+## 3. Recommended Improvements
+
+### Priority 1: Restore Build & Test Infrastructure
+1. **Implement Vite:** Restore/Create `vite.config.js` to enable bundling, minification, and modern development features.
+2. **Implement Vitest:** Restore the testing framework and re-establish the unit tests for core utilities and components.
+3. **Update `package.json`:** Re-add `build`, `preview`, and `test` scripts.
+
+### Priority 2: Architectural Refactoring
+1. **Refactor `main.js`**:
+   - Extract console filtering to `core/console-filter.js`.
+   - Extract global event delegation (sharing, retrying) to `core/global-events.js`.
+   - Move persistent storage logic to `core/pwa-manager.js`.
+2. **Modernize Resize Handling**: Replace `window.resize` listeners with `ResizeObserver` for better performance and element-relative reactivity.
+3. **Critical CSS Inlining**: Inline essential styles (Loader, Hero) in `index.html` to eliminate FOUC.
+
+### Priority 3: Asset Optimization
+1. **Build-time Image Checks**: Replace runtime `HEAD` requests for checking image existence with a build-time manifest or automated generation.
+
+## 4. Conclusion
+
+The `iweb` repository is a robust foundation with excellent code quality. By reconciling the current state with the documented infrastructure (Vite/Vitest) and applying the suggested refactorings, the project will achieve its full potential in terms of maintainability and performance.
+
+---
+*Report generated by Jules.*
